plugins/gmail/skills/gmail/pyproject.toml
[project]
name = "gmail-skill"
version = "0.1.0"
description = "Gmail sync skill for Claude Code"
requires-python = ">=3.11"
dependencies = [
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "httplib2>=0.22.0",
    "pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
